update source from base

+++ OgreMain/include/OgreEntity.h
@@ -37,6 +37,7 @@ THE SOFTWARE.
 #include "OgreHardwareBufferManager.h"
 #include "OgreRenderable.h"
 #include "OgreResourceGroupManager.h"
+#include "OgreHeaderPrefix.h"
 
 namespace Ogre {
     /** \addtogroup Core
@@ -366,7 +367,7 @@ namespace Ogre {
         /** Gets a pointer to a SubEntity, ie a part of an Entity.
          @deprecated use getSubEntities()
         */
-        SubEntity* getSubEntity(unsigned int index) const;
+        SubEntity* getSubEntity(size_t index) const;
 
         /** Gets a pointer to a SubEntity by name
         @remarks 
@@ -377,7 +378,7 @@ namespace Ogre {
         /** Retrieves the number of SubEntity objects making up this entity.
         * @deprecated use getSubEntities()
         */
-        unsigned int getNumSubEntities(void) const;
+        size_t getNumSubEntities(void) const;
 
         /** Retrieves SubEntity objects making up this entity.
         */
@@ -914,5 +915,6 @@ namespace Ogre {
 
 } // namespace Ogre
 
+#include "OgreHeaderSuffix.h"
 
 #endif // __Entity_H__
